Update BestGuess to also factor group counts
The BestGuess script previously only looked at the "minimax" of each word - finding the word for which the largest group of possible answers after getting any particular result was the smallest. This is still in place, however previously ties were broken using the first word that appears in the list of possible answers. Now it also checks that the number of different groups is the largest among words with the same minimax before tiebreaking as before.

I also restored the default starting word to the code's choice of ARISE over the WordleBot choice of CRANE.